Validate that flags exist
The current is_flag_enabled / is_flag_disabled macros don't
check that the flag is actually visible to policy before reading
the value. If, for example, a flag is referenced prior to being
added to flagging/Android.bp, the macros will silently work even
though the actual flag value isn't read.

Add error checks to the is_flag_enabled / is_flag_disabled macro
to assert that the flag is visible to policy. This will generate
a compiler failure when the flag isn't visible.

This isn't just a hypothetical situation. See b/410072692 starting
at comment #4, or b/442711782 as an example of the kind of errors
this change will detect.

This is the third attempt to submit this patch. The first attempt
was reverted due to b/442711782, and the second attempt was
reverted due to b/442711782#comment11. These bugs are fixed and
hopefully no further issues remain. There has been no change in
this patch since initial submission.

Flag: EXEMPT compile time change only
Test: Add a reference to an undefined flag and generate a compiler error
Test: see b/410072692 starting from comment #4
Test: see b/442711782
Change-Id: I6d0b96b6843b7fea850a1841830acfb91ff591df