Cleanup ashmem-memfd compatibility ioctl usage
Currently, any memfd can have an ashmem ioctl command invoked on it,
and it will work. The intent of this was to allow legacy applications
to be able to blindly assume a given fd is an ashmem fd.

The current rules are overly permissive, and can introduce unwanted
dependencies on this compatibility layer that will make it difficult
to deprecate in the future.

So, restrict the ashmem ioctl commands to work on memfds created by
libcutils--and thus labeled as ashmem_compat_memfd--for legacy
applications that might expect ashmem fds.

This will make it so that when these legacy applications are no longer
used, the compatibility layer can be removed.

It should be okay to remove ioctl usage for appdomain_tmpfs and tmpfs
domains since this was only introduced to allow the libcutils code to
check if the kernel supported the ashmem ioctls on memfds. Since
we're checking the vendor API level in libcutils instead to ensure
that the kernel is new enough, removing those permissions shouldn't
be an issue.

Bug: 443130838
Change-Id: Id1ebef193cd4747736ef3a94d6ef399eccc80243
Signed-off-by: Isaac J. Manjarres <isaacmanjarres@google.com>