Enable netlink_xperm capability
The new netlink_xperm capability is enabled for kernel >= 6.13. This
ensures that the new extended permission (nlmsg) is used in place of the
previous permissions (i.e., nlmsg_read, nlmsg_write, nlmsg_readpriv and
nlmsg_getneigh).

On older kernels, this unknown capability is simply ignored.

This is a re-apply of commit fa4957fbb25a.

Bug: 353255679
Test: presubmit
Flag: EXEMPT policy change
Change-Id: I97cfd6967c0beb64d1b1019508e13aa31150d945