allow fastboot HALs to access the misc partition
Currently, fastbootd is allowed to access the misc_block_device, but
this is not true for hal_fastboot_server. This makes it impossible to
implement vendor-specific fastboot OEM commands that modify the misc
partition (e.g. to manipulate data structures stored in the vendor space
of the misc partition), because OEM commands are handled by
the fastboot HAL. Allow hal_fastboot_server to access
the misc_block_device.

Bug: 407891018
Test: `allow hal_fastboot_default misc_block_device:blk_file read`
sepolicy rule doesn't result in build failure.

Change-Id: I0c42f3cdf8662d1fc0d9a506b1ce9353c9d7dd55