Clean up the pmgd service sepolicy
Rename "pmg" to "pmgd" (for Process Memory Guardian Daemon) to be
consistent with the user and process name.

The pmgd service now exists and and is enabled if the build flag is
enabled.  The rc init script no longer relies on a sysprop so the
corresponding selinux permissions are removed.

Tested with the aconfig flag enabled and disabled.  In both cases the
daemon is started once by init.  The operational state of the daemon
is determined from the log messages.

Verified that no selinux denials related to pmgd were observed.

Flag: EXEMPT UNSUPPORTED_LANGUAGE
Bug: 434681786
Test: presubmit
Change-Id: I0cde79078b33232f740187c2d136575466cb224d