SePolicy that introduces pcc_component domain
The new domain is derived from ephemeral_app. In this CL all of the
allow and neverallow rules of ephemeral_app have been copied over with
a few notable exceptions:
- Allow read access to storage_file and mnt_user_file dir, file.
- Allow pcc_commponent to search/open files/dirs of type sdcard_type, fuse,
  media_rw_data_file.
- Allow keystore access.
- Allow omapi access.
- All the ephemeral_app_api_service services can be used by
  pcc_component but we're not reusing the type attribute to define this
  rule. Instead we use an explicit allowlist for better maintainability.

This is the first of many CLs that set up this domain. In the follow up
CLs this domain will further diverge from ephemeral_app.

Bug: 430201041
Test: m, push sepolicy and boot device
Flag: N/A
Change-Id: I66e9bbff60e24f7d929e329f7c5333b2086c5ae3