Configure a domain for mmd_setup
Zram setup was previously done by passing an argument to mmd and
directly exec'ing it into the su domain. This was done to avoid giving
unnecessary elevated privileges to the long-running mmd service.
However, that approach doesn't work on user builds, since the su domain
isn't present. To fix this, the mmd binary is being split into two parts
- one for setup and one for the service - each with their own domain.

This change defines the new mmd_setup domain. It is fairly privileged
since it needs to enable swap and set up the backing device for zram
writeback.

Bug: 450175598
Test: manually verify mmd starts on a user build
Change-Id: Ie7ade44f07b42a558a69c578d62a9deb7ecee836