Add SELinux permissions for traced.
This change allows `traced` to create and write to files within
`/data/misc/perfetto-traces/bugreport*`.

Previously only `perfetto` was allowed to create and to write to
`bugreport*` files (
https://source.corp.google.com/h/googleplex-android/platform/superproject/main/+/main:system/sepolicy/private/perfetto.te;l=26-27;drc=c5401edfb4d65686275a3471e38cefbc1913a70f
).

This additional permission is required for us to merge a new way of
cloning the long (`write_into_file`) traces, that is also how we create
a perfetto bugreports on Android. With this new logic in some cases
`traced` itself write the bugreports instead of sending data to the
perfetto process.

Before this SELinux change bugreport creation failed with the following
avc error message:

```
10-07 16:12:25.908  1200  1200 I auditd  : type=1400 audit(0.0:54): avc:  denied  { read write } for  comm="traced" path="/data/misc/perfetto-traces/bugreport/systrace.pftrace" dev="dm-78" ino=7377 scontext=u:r:traced:s0 tcontext=u:object_r:perfetto_traces_bugreport_data_file:s0 tclass=file permissive=0
```

with this change everything works fine.

See
https://github.com/google/perfetto/commit/b5a3a4b38331ca4b3362a664451799b437fcc5f5
for the implementation details (or see the CL attached to this topic).

Bug: b/382209797
Test: atest CtsPerfettoTestCases

Change-Id: I3688964826c85241089b6ea222796db11211990b