Let omapi register secure_element_service
Move omapi type to public policy so vendor components can be the subject
of the `use_omapi()` macro.

Introduce an attribute for domains that implement OMAPI via the
secure_element_service services.

Remove the omapi_service type that was never associated to an actual
service and assume it was meant to refer to the secure_element_service
APIs.

Test: Boot on watch and phone build
Fix: 447424308
Bug: 449793480
Flag: EXEMPT BUGFIX
Change-Id: I475ea17ed07eb52c3849945d64c23936d7c3d4b2