Undo memfd_file addition to constraints
I erroneously added memfd_file to the set of objects that should
be constrained for open permissions. However, this is unreachable,
as the thought was that a memfd could be opened by another process
via /proc/$pid/fd/$fd. However, since those files are classified as
lnk_file, and open isn't checked for those files, the policy is not
reachable, so remove it.

Bug: 404959459
Bug: 111903542
Change-Id: Ia788fbd8303785fe446d7d26c19dd1b447df2733
Signed-off-by: Isaac J. Manjarres <isaacmanjarres@google.com>