Revert "Create untrusted app domain for target SDK level 36"
This reverts commit 7c1f5bae67148051305545622655f0cc51f75009.

Splitting up the domains should not affect 26Q1 so lets revert
this for now and wait until 26Q2.

Note: Since the expected target SDK version is 36 at the moment,
and it will be the last one to be able to invoke ashmem ioctl
commands with memfd, untrusted_app.te will have that permission.
This will later be removed when untrusted_app becomes the domain
for untrusted apps targeting SDK version 37+.

Bug: 456406505
Bug: 455888842
Bug: 455887079
Bug: 443130838
Change-Id: I82555b890d0771a19e5fc9b65412a434bc6d2b00
Signed-off-by: Isaac J. Manjarres <isaacmanjarres@google.com>