Revert "Split privileged app domain based on target SDK level"
This reverts commit 9ea40f63d04a83337567ac3284063f2d7cb3df72.

Splitting up the domains should not affect 26Q1 so lets back
this out for now and wait until 26Q2.

Note: Since the expected target SDK version is 36 at the moment,
and it will be the last one to be able to invoke ashmem ioctl
commands with memfd, priv_app.te will have that permission.
This will later be removed when priv_app becomes the domain
for privileged apps targeting SDK version 37+.

Bug: 456406505
Bug: 455888842
Bug: 455887079
Bug: 443130838
Change-Id: I761c306f1ab4bd5b360d0986ef72d887ca86fd23
Signed-off-by: Isaac J. Manjarres <isaacmanjarres@google.com>