Reland: sepolicy: Allow bpfloader to attach tracepoint BPF programs
Grant the bpfloader permissions to read tracepoint event information
from tracefs.

To attach a BPF program to a tracepoint, the loader must read the
trace event's ID from the tracefs virtual filesystem. This change
adds the necessary allow rule to access files labeled
debugfs_tracing_debug, which corresponds to tracefs.

An exception to a neverallow rule is also added, making this a
required change for loading any tracepoint-based BPF programs.

Test: boot; no avc denials
Bug: 442017923
Change-Id: I09daec727f8387ac2340cdfcfee82ea7105e88f2
Signed-off-by: Kalesh Singh <kaleshsingh@google.com>