Allow init_dev_config to set system_vendor_config_prop() properties
The 'init_dev_config' service does some of the work that 'vendor_init'
would normally do, such as initializing system properties whose values
are determined by the vendor. 'vendor_init' does this by reading from
the *.prop files, and 'init_dev_config'  does this more dynamically,
but they have essentially the same purpose.

The 'system_vendor_config_prop' macro is updated to make sure that the
process in the 'init_dev_config' domain is allowed to set these
properties, in the same way as a process in the 'vendor_init' domain
would.

Bug: 453917315
Test: m selinux_policy

Change-Id: If2bc0740ab07e04315f5fe6b0c8414116fa77ce9