Restore shell's ability to list block devices
http://ag/36574287 ("Add sysfs_block type for mmd_setup") relabeled
/sys/class/block from sysfs to sysfs_block.  That caused shell to no
longer be able to traverse this directory.  In turn, that caused
CtsNativeEncryptionTestCases FileBasedEncryptionPolicyTest#allowedPolicy
to start failing on devices that use the emmc_optimized encryption
option, because it relies on lstat("/sys/class/block/mmcblk0") working.

Fix this by granting shell the same permission on sysfs_block:dir that
it has on sysfs:dir.

Bug: 463597109
Test: atest CtsNativeEncryptionTestCases # on device using emmc_optimized
Flag: EXEMPT BUGFIX
Change-Id: Id197eb716cfcc69580017686880d68212fd24a98